feat: top snaps preview
