+++ snaprecommend/templates/layout.html
@@ -74,12 +74,8 @@
   </header>
 
   <main class="l-main">
-    <div class="p-panel">
-      <div class="u-fixed-width">
         {% block page_content %}
         {% endblock %}
-      </div>
-    </div>
   </main>
 </div>
 {% endblock %}
