+++ snaprecommend/templates/partials/snap_card.html
@@ -1,0 +1,9 @@
+<li>
+    <div class="p-card">
+        <div style="display:flex; gap:12px;">
+            <img src="{{ snap.icon }}" alt="{{ snap.name }}" style="height:50px;" class="u-icon">
+            <h4><a href="https://snapcraft.io/{{ snap.name }}" target="_blank">{{ snap.name }}</a></h4>
+        </div>
+        <p class="card__content u-truncate">{{ snap.summary }}</p>
+    </div>
+</li>
